requirements.txt
PyGithub==2.1.1
python-dotenv==1.0.0
fastmcp==0.1.0 